Refactor #1 Change Amount Long to Double - 국내의 경우 금액에 대해 Long, Int를 쓰나 미국과 같이 소숫점을 쓰는 경우도 존재하기 때문에 시스템은 기본적으로 Double을 쓰고 이를 Paygate module 내부적으로 타입 변경을 하도록 한다. - 통화의 경우 Product나 CP에 영향을 받기 보다 사용자의 결제수단에 영향을 받기 때문에 PrepareInteraction.kt을 할 때 통화 정보를 받아 적절한 결제 가능 수단을 노출 하도록 한다.

+++ paygate/src/main/kotlin/io/mustelidae/weasel/paygate/domain/client/inicis/InicisDummyClient.kt
@@ -31,7 +31,7 @@ internal class InicisDummyClient(
             LocalDateTime.now(),
             request.storeId,
             "Weasel",
-            request.token.paymentAmount
+            request.token.paymentAmount.toLong()
         ).apply {
             credit = InicisResources.Reply.Paid.Credit(
                 CreditCode.AMEX.inicis,
@@ -40,7 +40,7 @@ internal class InicisDummyClient(
                 isInterestFree = false,
                 isCheckCard = true,
                 interestMonth = "10",
-                applyPrice = request.token.paymentAmount,
+                applyPrice = request.token.paymentAmount.toLong(),
                 approveCode = "1235143",
                 name = CreditCode.AMEX.name,
                 issuer = null,
